lambda/fetch_data.py
import os
import boto3
import requests

def handler(event, context):
    api_key = os.environ.get('API_KEY')
    bucket = os.environ.get('BUCKET_NAME')
    
    # Placeholder for Alpha Vantage fetch logic
    print("Fetching data from Alpha Vantage...")
    
    return {
        'statusCode': 200,
        'body': 'Data fetched and stored'
    }
